implemented user tests

+++ routes/web.php
@@ -22,14 +22,14 @@ Route::get('/', [TripController::class,'index'])->name('home');
 Route::get('/home', [TripController::class,'index']);
 
 //D del CRUD
-Route::delete('/delete/{id}', [TripController::class, 'destroy'])->name('deleteTrip');
+Route::delete('/delete/{id}', [TripController::class, 'destroy'])->name('deleteTrip')->middleware('isadmin', 'auth');
 
 //C del CRUD
-Route::get('/create', [TripController::class, 'create'])->name('createTrip');
-Route::post('/', [TripController::class, 'store'])->name('storeTrip');
+Route::get('/create', [TripController::class, 'create'])->name('createTrip')->middleware('isadmin', 'auth');
+Route::post('/', [TripController::class, 'store'])->name('storeTrip')->middleware('isadmin', 'auth');
 
 //U del CRUD
-Route::get('/edit/{id}', [TripController::class, 'edit'])->name('editTrip');
-Route::patch('/trip/{id}', [TripController::class, 'update'])->name('updateTrip');
+Route::get('/edit/{id}', [TripController::class, 'edit'])->name('editTrip')->middleware('isadmin', 'auth');
+Route::patch('/trip/{id}', [TripController::class, 'update'])->name('updateTrip')->middleware('isadmin', 'auth');
 
 Route::get('/show/{id}', [TripController::class, 'show'])->name('showTrip');
